完成 student 接口 (#40)
* Refactor GET /student/search

* Refactor POST /student & PUT /student

* Refactor PUT /student/grade & POST /student/grade

* Refactor POST /student/apply

* Refactor GET /student

+++ src/main/java/backend/model/VO/student/Score.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class Score {
-    private Float gradeTotal;
+    private Double gradeTotal;
 
     private List<GradeList> gradeList;
 }
